backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
geoalchemy2==0.15.1
psycopg2-binary==2.9.9
httpx==0.27.0
pydantic==2.7.4
anthropic==0.28.0
xgboost==2.0.3
numpy==1.26.4
shapely==2.0.4
python-dotenv==1.0.1
asyncache==0.3.1
cachetools==5.3.3
sse-starlette==2.1.0
